+++ ModLoader/ModEntry.cs
@@ -186,7 +186,7 @@ namespace Sheepy.Modnix {
          switch ( id ) {
             case "mods_root" : return ModLoader.ModDirectory;
             case "loader" : case "modnix" :
-               return Assembly.GetExecutingAssembly().Location;
+               return ModLoader.LoaderPath;
             case "phoenixpoint" : case "phoenix point" : case "game" :
                return Process.GetCurrentProcess().MainModule?.FileName;
             default :
@@ -282,9 +282,9 @@ namespace Sheepy.Modnix {
       #endregion
 
       #region Logger
-      internal LoggerProxy Logger; // Created when and only when an initialiser accepts a logging function
+      private LoggerProxy Logger; // Created when and only when an initialiser accepts a logging function
 
-      private Logger CreateLogger () {
+      public Logger CreateLogger () {
          lock ( this ) {
             if ( Logger != null ) return Logger;
             Logger = new LoggerProxy( ModLoader.Log ){ Level = LogLevel ?? ModLoader.Settings.LogLevel };
@@ -515,6 +515,7 @@ namespace Sheepy.Modnix {
    public class ModMeta {
       public string   Id;
       public Version  Version;
+      public string[]  Flags;
 
       public TextSet   Name;
       public string[]  Lang;
@@ -531,19 +532,21 @@ namespace Sheepy.Modnix {
 
       public string[]  Mods;
       public DllMeta[] Dlls;
+      public object Actions;
 
       public   string  ConfigType;
       public   object  DefaultConfig;
       internal string  DefaultConfigText;
       internal string  ConfigText;
 
-      internal bool HasContent => Mods == null && Dlls == null;
+      internal bool HasContent => Mods == null && Dlls == null && Actions == null;
 
       internal ModMeta ImportFrom ( ModMeta overrider ) {
          lock ( this ) if ( overrider == null ) return this;
          lock ( overrider ) {
             CopyNonNull( overrider.Id, ref Id );
             CopyNonNull( overrider.Version, ref Version );
+            CopyNonNull( overrider.Flags, ref Flags );
             CopyNonNull( overrider.Name, ref Name );
             CopyNonNull( overrider.Lang, ref Lang );
             CopyNonNull( overrider.Duration, ref Duration );
@@ -557,6 +560,7 @@ namespace Sheepy.Modnix {
             CopyNonNull( overrider.LoadIndex, ref LoadIndex );
             CopyNonNull( overrider.Mods, ref Mods );
             CopyNonNull( overrider.Dlls, ref Dlls );
+            CopyNonNull( overrider.Actions, ref Actions );
             CopyNonNull( overrider.ConfigType, ref ConfigType );
             CopyNonNull( overrider.DefaultConfig, ref DefaultConfig );
          }
@@ -579,6 +583,7 @@ namespace Sheepy.Modnix {
          NormTextSet( ref Name );
          if ( Name == null && Id != null )
             Name = new TextSet{ Default = Id };
+         NormStringArray( ref Flags );
          NormStringArray( ref Lang );
          Duration = NormString( Duration );
          NormTextSet( ref Description );